Rename rtlPalWithContexts to AppTester.
