+++ frontend/lib/tests/pages/onboarding-step-4.test.tsx
@@ -1,14 +1,14 @@
 import React from 'react';
 
 import OnboardingStep4 from '../../pages/onboarding-step-4';
-import { rtlPalWithContexts } from '../rtl-pal-with-contexts';
+import { AppTester } from '../app-tester';
 
 
 describe('onboarding step 2 page', () => {
-  afterEach(rtlPalWithContexts.cleanup);
+  afterEach(AppTester.cleanup);
 
   it('displays welcome modal on successful signup', async () => {
-    const { pal, client } = rtlPalWithContexts(<OnboardingStep4 />);
+    const { pal, client } = new AppTester(<OnboardingStep4 />);
 
     pal.clickButtonOrLink("Create account");
     client.getRequestQueue()[0].resolve({ output: { errors: [], session: {} } });
